+++ lib/flowcommerce_spree.rb
@@ -6,7 +6,6 @@ require 'flowcommerce_spree/refresher'
 require 'flowcommerce_spree/engine'
 require 'flowcommerce_spree/logging_http_client'
 require 'flowcommerce_spree/logging_http_handler'
-require 'flowcommerce_spree/webhook_service'
 require 'flowcommerce_spree/session'
 require 'flow/simple_gateway'
