+++ spec/dummy/config/environments/test.rb
@@ -12,7 +12,7 @@ Rails.application.configure do
   # Do not eager load code on boot. This avoids loading your whole application
   # just for the purpose of running a single test. If you are using a tool that
   # preloads Rails for running tests, you may have to set it to true.
-  config.eager_load = false
+  config.eager_load = true
 
   config.action_controller.default_url_options = { host: 'test.host' }
   routes.default_url_options = config.action_controller.default_url_options
